Replaced the ClosableTabButtonSkin with IconClosableTabButtonSkin.
The new skin supports displaying:

An icon to the left or right of the text,

Text without an icon, or

An icon without text.

The nested close button functionality remains intact.

This new skin has been integrated into the CustomButtonBarSkin master skin, so changes should be fully transparent.

To use the new feature, specify a content key in the dataset for any item in the TabBar data provider. Set content to one of the following:

A String,
A SpriteVisualElement instance, or
An Array containing any combination of these (including a single item).

You only need to use the content key where required. The label property will continue to work but will be overridden when content is specified.